Add fallback to English locale if not value is given in the config
The localisation feature requires an additional entry in the config, as
has been stated in the last commit.
To not break existing configs and retain backwards-compatibility, this
commit selects English as default locale if no entry is put in the
config and notifies the user through the logs about the missing entry.